added share feature

+++ git/models.py
@@ -7,7 +7,7 @@ class Chunk(models.Model):
     fileId=models.ForeignKey(FileId,on_delete=models.CASCADE)
     repo=models.CharField(max_length=12)
     uname=models.CharField(max_length=12)
-    size = models.FloatField()
+    size = models.IntegerField()
     
 class RepoSize(models.Model):
     name = models.CharField(max_length=255, unique=True)
